refactor: rebase utils functions structure
