+++ dapp/src/components/ConnectWallet.astro
@@ -3,7 +3,7 @@
 </div>
 
 <script>
-  import { truncateMiddle } from "../service/utils";
+  import { truncateMiddle } from "../utils/utils";
   import { kit, loadedPublicKey, setPublicKey } from "./stellar-wallets-kit";
 
   document.addEventListener("astro:page-load", () => {
