+++ dapp/src/utils/editLinkFunctions.ts
@@ -10,3 +10,29 @@ export function convertGitHubLink(link: string): string {
     return link;
   }
 }
+
+export function getGithubContentUrl(username: string, repoName: string, filePath: string): string {
+  return `https://raw.githubusercontent.com/${username}/${repoName}/${filePath}`;
+}
+
+export function getGithubContentLink(username: string, repoName: string, filePath: string): string {
+  return `https://github.com/${username}/${repoName}/blob/${filePath}`;
+}
+
+export function getAuthorRepo(repoUrl: string): { username: string | undefined; repoName: string | undefined } {
+  const match = repoUrl.match(/https\:\/\/github\.com\/([^\/]+)\/([^\/]+)/);
+  if (!match || !match[1] || !match[2]) return {username: undefined, repoName: undefined};
+  return { username: match[1], repoName: match[2] };
+}
+
+export function getGithubContentUrlFromConfigUrl(configUrl: string): string | undefined {
+  const { username, repoName } = getAuthorRepo(configUrl);
+  if (username && repoName) {
+    return getGithubContentUrl(username, repoName, 'main/tansu.toml');
+  }
+  return undefined;
+}
+
+export function getGithubContentLinkFromConfigUrl(configUrl: string, filePath: string): string {
+  return `${configUrl}/blob/${filePath}`;
+}
