+++ dapp/src/utils/formatTimeFunctions.ts
@@ -1,0 +1,23 @@
+export function formatTime(dateString: string): string {
+  const date = new Date(dateString);
+  const today = new Date();
+  
+  if (
+    date.getDate() === today.getDate() &&
+    date.getMonth() === today.getMonth() &&
+    date.getFullYear() === today.getFullYear()
+  ) {
+    const timeOptions: Intl.DateTimeFormatOptions = { hour: '2-digit', minute: '2-digit' };
+    const formattedTime = date.toLocaleTimeString(undefined, timeOptions);
+    return `Today, ${formattedTime}`;
+  } else {
+    const dateOptions: Intl.DateTimeFormatOptions = { year: 'numeric', month: 'short', day: 'numeric' };
+    return date.toLocaleDateString(undefined, dateOptions);
+  }
+}
+
+export function formatDate(dateString: string): string {
+  const date = new Date(dateString);
+  const dateOptions: Intl.DateTimeFormatOptions = { year: 'numeric', month: 'short', day: 'numeric' };
+  return date.toLocaleDateString(undefined, dateOptions);
+}
