+++ dapp/src/utils/utils.ts
@@ -1,0 +1,9 @@
+export function truncateMiddle(str: string, maxLength: number): string {
+  if (str.length <= maxLength) return str;
+  const ellipsis = '...';
+  const charsToShow = maxLength - ellipsis.length;
+  const frontChars = Math.ceil(charsToShow / 2);
+  const backChars = Math.floor(charsToShow / 2);
+  return str.substr(0, frontChars) + ellipsis + str.substr(str.length - backChars);
+}
+
